requirements.txt
Django==5.1.3
django-jazzmin==3.0.1
Pillow==11.0.0
django-recaptcha==4.0.0
django-colorfield==0.11.0
django-admin-sortable2==2.2.4
django-ckeditor-5==0.2.15
django-compressor==4.5.1
beautifulsoup4==4.12.3